+++ src/egon/data/datasets/industry/__init__.py
@@ -7,18 +7,17 @@
 """
 
 
-import egon.data.config
+from sqlalchemy import ARRAY, Column, Float, Integer, String
+from sqlalchemy.ext.declarative import declarative_base
 import geopandas as gpd
-import pandas as pd
+
 from egon.data import db
 from egon.data.datasets import Dataset
 from egon.data.datasets.industry.temporal import (
     insert_osm_ind_load,
     insert_sites_ind_load,
 )
-from sqlalchemy import Column, String, Float, Integer, ARRAY
-from sqlalchemy.ext.declarative import declarative_base
-
+import egon.data.config
 
 Base = declarative_base()
 
@@ -51,6 +50,19 @@ class DemandCurvesOsmIndustry(Base):
     p_set = Column(ARRAY(Float))
 
 
+class DemandCurvesOsmIndustryIndividual(Base):
+    __tablename__ = "egon_osm_ind_load_curves_individual"
+    __table_args__ = {"schema": "demand"}
+
+    osm_id = Column(Integer, primary_key=True)
+    bus_id = Column(Integer)
+    scn_name = Column(String, primary_key=True)
+    p_set = Column(ARRAY(Float))
+    peak_load = Column(Float)
+    demand = Column(Float)
+    voltage_level = Column(Integer)
+
+
 class DemandCurvesSitesIndustry(Base):
     __tablename__ = "egon_sites_ind_load_curves"
     __table_args__ = {"schema": "demand"}
@@ -59,6 +71,20 @@ class DemandCurvesSitesIndustry(Base):
     scn_name = Column(String, primary_key=True)
     wz = Column(Integer, primary_key=True)
     p_set = Column(ARRAY(Float))
+
+
+class DemandCurvesSitesIndustryIndividual(Base):
+    __tablename__ = "egon_sites_ind_load_curves_individual"
+    __table_args__ = {"schema": "demand"}
+
+    site_id = Column(Integer, primary_key=True)
+    bus_id = Column(Integer)
+    scn_name = Column(String, primary_key=True)
+    p_set = Column(ARRAY(Float))
+    peak_load = Column(Float)
+    demand = Column(Float)
+    voltage_level = Column(Integer)
+    wz = Column(Integer)
 
 
 def create_tables():
@@ -95,12 +121,26 @@ def create_tables():
 
     db.execute_sql(
         f"""DROP TABLE IF EXISTS
-            {targets_temporal['osm_load']['schema']}.{targets_temporal['osm_load']['table']} CASCADE;"""
+            {targets_temporal['osm_load']['schema']}.
+            {targets_temporal['osm_load']['table']} CASCADE;"""
     )
 
     db.execute_sql(
         f"""DROP TABLE IF EXISTS
-            {targets_temporal['sites_load']['schema']}.{targets_temporal['sites_load']['table']} CASCADE;"""
+            {targets_temporal['osm_load_individual']['schema']}.
+            {targets_temporal['osm_load_individual']['table']} CASCADE;"""
+    )
+
+    db.execute_sql(
+        f"""DROP TABLE IF EXISTS
+            {targets_temporal['sites_load']['schema']}.
+            {targets_temporal['sites_load']['table']} CASCADE;"""
+    )
+
+    db.execute_sql(
+        f"""DROP TABLE IF EXISTS
+            {targets_temporal['sites_load_individual']['schema']}.
+            {targets_temporal['sites_load_individual']['table']} CASCADE;"""
     )
 
     engine = db.engine()
@@ -115,11 +155,19 @@ def create_tables():
 
     DemandCurvesOsmIndustry.__table__.create(bind=engine, checkfirst=True)
 
+    DemandCurvesOsmIndustryIndividual.__table__.create(
+        bind=engine, checkfirst=True
+    )
+
     DemandCurvesSitesIndustry.__table__.create(bind=engine, checkfirst=True)
+
+    DemandCurvesSitesIndustryIndividual.__table__.create(
+        bind=engine, checkfirst=True
+    )
 
 
 def industrial_demand_distr():
-    """ Distribute electrical demands for industry to osm landuse polygons
+    """Distribute electrical demands for industry to osm landuse polygons
     and/or industrial sites, identified earlier in the process.
     The demands per subsector on nuts3-level from demandregio are distributed
     linearly to the area of the corresponding landuse polygons or evenly to
@@ -172,7 +220,9 @@ def industrial_demand_distr():
                 WHERE sector = 3
                 AND NOT ST_Intersects(
                     geom,
-                    (SELECT ST_UNION(ST_Transform(geom,3035)) FROM {sources['industrial_sites']['schema']}.{sources['industrial_sites']['table']}))
+                    (SELECT ST_UNION(ST_Transform(geom,3035)) FROM
+                    {sources['industrial_sites']['schema']}.
+                    {sources['industrial_sites']['table']}))
                 AND name NOT LIKE '%%kraftwerk%%'
                 AND name NOT LIKE '%%Stadtwerke%%'
                 AND name NOT LIKE '%%Müllverbrennung%%'
@@ -210,7 +260,8 @@ def industrial_demand_distr():
                 {sources['industrial_sites']['table']}""",
             index_col=None,
         )
-        # Count number of industrial sites per subsector (wz) and nuts3 district
+        # Count number of industrial sites per subsector (wz) and nuts3
+        # district
         sites_grouped = (
             sites.groupby(["nuts3", "wz"]).size().reset_index(name="counts")
         )
@@ -227,30 +278,36 @@ def industrial_demand_distr():
                          WHERE sector = 'industry')"""
         )
 
-        # Replace wz=17 and wz=18 by wz=1718 as a differentiation of these two subsectors can't be performed
+        # Replace wz=17 and wz=18 by wz=1718 as a differentiation of these two
+        # subsectors can't be performed
         demand_nuts3_import["wz"] = demand_nuts3_import["wz"].replace(
             [17, 18], 1718
         )
 
-        # Group results by nuts3 and wz to aggregate demands from subsectors 17 and 18
+        # Group results by nuts3 and wz to aggregate demands from subsectors
+        # 17 and 18
         demand_nuts3 = (
             demand_nuts3_import.groupby(["nuts3", "wz"]).sum().reset_index()
         )
 
-        # A differentiation between those industrial subsectors (wz) which aren't represented
-        # and subsectors with a representation in the data set on industrial sites is needed
+        # A differentiation between those industrial subsectors (wz) which
+        # aren't represented and subsectors with a representation in the
+        # dataset on industrial sites is needed
 
-        # Select industrial demand for sectors which aren't found in industrial sites as category a
+        # Select industrial demand for sectors which aren't found in
+        # industrial sites as category a
         demand_nuts3_a = demand_nuts3[
             ~demand_nuts3["wz"].isin([1718, 19, 20, 23, 24])
         ]
 
-        # Select industrial demand for sectors which are found in industrial sites as category b
+        # Select industrial demand for sectors which are found in industrial
+        # sites as category b
         demand_nuts3_b = demand_nuts3[
             demand_nuts3["wz"].isin([1718, 19, 20, 23, 24])
         ]
 
-        # Bring demands on nuts3 level and information on industrial sites per nuts3 district together
+        # Bring demands on nuts3 level and information on industrial sites per
+        # nuts3 district together
         demand_nuts3_b = demand_nuts3_b.merge(
             sites_grouped,
             how="left",
@@ -258,7 +315,8 @@ def industrial_demand_distr():
             right_on=["nuts3", "wz"],
         )
 
-        # Define share of industrial demand per nuts3 region and subsector allocated to industrial sites
+        # Define share of industrial demand per nuts3 region and subsector
+        # allocated to industrial sites
         share_to_sites = 0.5
 
         # Define demand per site for every nuts3 region and subsector
@@ -270,7 +328,8 @@ def industrial_demand_distr():
         # Replace NaN by 0
         demand_nuts3_b = demand_nuts3_b.fillna(0)
 
-        # Calculate demand which needs to be distributed to osm landuse areas from category b
+        # Calculate demand which needs to be distributed to osm landuse areas
+        # from category b
         demand_nuts3_b["demand_b_osm"] = demand_nuts3_b["demand"] - (
             demand_nuts3_b["demand_per_site"] * demand_nuts3_b["counts"]
         )
@@ -290,7 +349,8 @@ def industrial_demand_distr():
             {"demand_b_osm": "demand"}, axis=1
         )
 
-        # Create df containing all demand per wz which will be allocated to osm areas
+        # Create df containing all demand per wz which will be allocated to
+        # osm areas
         demand_nuts3_osm_wz = demand_nuts3_a.append(
             demand_nuts3_b_osm, ignore_index=True
         )
@@ -350,7 +410,7 @@ class IndustrialDemandCurves(Dataset):
     def __init__(self, dependencies):
         super().__init__(
             name="Industrial_demand_curves",
-            version="0.0.4",
+            version="0.0.5",
             dependencies=dependencies,
             tasks=(
                 create_tables,