+++ src/egon/data/datasets/scenario_parameters/parameters.py
@@ -833,6 +833,9 @@ def heat(scenario):
                 costs, "central resistive heater", "VOM"
             ),
             "geo_thermal": 2.9,  # Danish Energy Agency
+            "water_tank_charger": 0,  # Danish Energy Agency
+            "water_tank_discharger": 0,  # Danish Energy Agency
+            "rural_heat_pump": 0,  # Danish Energy Agency, Technology Data for Individual Heating Plants
         }
 
     elif scenario == "eGon100RE":